+++ mini-projects/day-06-dict.py
@@ -1,0 +1,11 @@
+#Requirements
+#Ask the user for:
+#Name
+#Age
+#College
+#Course
+#Dream Company
+#sStore everything inside one dictionary.
+
+student = {"Name":"Adeeb","Favorite Food":"Biriyani","Age":20,"College":"CMR University","Course":"B.Tech CSE(Data Science)"}
+print(student)